+++ include/votca/tools/graph_df_visitor.h
@@ -43,7 +43,7 @@ class Graph_DF_Visitor : public GraphVisitor {
 
   /// The core of the breadth first visitor is in how the edges are added
   /// to the queue in this function
-  void addEdges_(const Graph& g, int vertex) override;
+  void addEdges_(const Graph& g, long int vertex) override;
   Edge getEdge_(const Graph& g) override;
 
  public:
