+++ include/votca/tools/reducedgraph.h
@@ -75,7 +75,7 @@ class ReducedGraph : public Graph {
              std::unordered_map<long int, GraphNode> nodes);
 
   // Junctions must be stored internally
-  std::set<int> junctions_;
+  std::set<long int> junctions_;
 
  public:
   ReducedGraph() = default;
@@ -130,9 +130,9 @@ class ReducedGraph : public Graph {
    *
    * Thus 1 is the only junction that exists in the reduced graph
    **/
-  // std::vector<int> getJunctions() const;
+  // std::vector<long int> getJunctions() const;
 
-  std::vector<std::pair<int, GraphNode>> getNodes(void) const override;
+  std::vector<std::pair<long int, GraphNode>> getNodes(void) const override;
 
   std::vector<long int> getVerticesDegree(long int degree) const override;
