+++ src/libtools/graph.cc
@@ -31,7 +31,7 @@ class GraphNode;
 
 bool nodeForEveryVertex_(vector<long int> vertices,
                          unordered_map<long int, GraphNode> nodes) {
-  for (const int vertex : vertices) {
+  for (const long int vertex : vertices) {
     if (nodes.count(vertex) == 0) {
       return false;
     }
@@ -45,7 +45,7 @@ Graph::Graph(vector<Edge> edges, unordered_map<long int, GraphNode> nodes) {
   assert(nodeForEveryVertex_(vertices, nodes) &&
          "A node must exist for every vertex.");
   nodes_ = nodes;
-  for (const pair<int, GraphNode> id_and_node : nodes_) {
+  for (const pair<long int, GraphNode> id_and_node : nodes_) {
     if (edge_container_.vertexExist(id_and_node.first) == false) {
       edge_container_.addVertex(id_and_node.first);
     }
@@ -59,54 +59,54 @@ bool Graph::operator!=(const Graph& graph) const {
 
 bool Graph::operator==(const Graph& graph) const { return !(*(this) != graph); }
 
-vector<pair<int, GraphNode>> Graph::getIsolatedNodes(void) const {
-  vector<pair<int, GraphNode>> isolated_nodes;
+vector<pair<long int, GraphNode>> Graph::getIsolatedNodes(void) const {
+  vector<pair<long int, GraphNode>> isolated_nodes;
   vector<long int> vertices_degree_0 = edge_container_.getVerticesDegree(0);
 
-  for (const int vertex : vertices_degree_0) {
-    pair<int, GraphNode> id_and_node{vertex, nodes_.at(vertex)};
+  for (const long int vertex : vertices_degree_0) {
+    pair<long int, GraphNode> id_and_node{vertex, nodes_.at(vertex)};
     isolated_nodes.push_back(id_and_node);
   }
   return isolated_nodes;
 }
 
-vector<pair<int, GraphNode>> Graph::getNeighNodes(int vertex) const {
-  vector<int> neigh_vertices = edge_container_.getNeighVertices(vertex);
-  vector<pair<int, GraphNode>> neigh_ids_and_nodes;
-  for (const int& neigh_vert : neigh_vertices) {
-    auto id_and_node = pair<int, GraphNode>(neigh_vert, nodes_.at(neigh_vert));
+vector<pair<long int, GraphNode>> Graph::getNeighNodes(long int vertex) const {
+  vector<long int> neigh_vertices = edge_container_.getNeighVertices(vertex);
+  vector<pair<long int, GraphNode>> neigh_ids_and_nodes;
+  for (const long int& neigh_vert : neigh_vertices) {
+    auto id_and_node = pair<long int, GraphNode>(neigh_vert, nodes_.at(neigh_vert));
     neigh_ids_and_nodes.push_back(id_and_node);
   }
   return neigh_ids_and_nodes;
 }
 
-void Graph::setNode(int vertex, GraphNode& graph_node) {
+void Graph::setNode(long int vertex, GraphNode& graph_node) {
   assert(nodes_.count(vertex) && "Can only set a node that already exists");
   nodes_[vertex] = graph_node;
   calcId_();
 }
 
-void Graph::setNode(std::pair<int, GraphNode>& id_and_node) {
+void Graph::setNode(std::pair<long int, GraphNode>& id_and_node) {
   setNode(id_and_node.first, id_and_node.second);
 }
 
-GraphNode Graph::getNode(const int vertex) const {
+GraphNode Graph::getNode(const long int vertex) const {
   assert(nodes_.count(vertex));
   return nodes_.at(vertex);
 }
 
-vector<pair<int, GraphNode>> Graph::getNodes(void) const {
-  vector<pair<int, GraphNode>> vec_nodes;
-  for (const pair<int, GraphNode>& id_and_node : nodes_) {
+vector<pair<long int, GraphNode>> Graph::getNodes(void) const {
+  vector<pair<long int, GraphNode>> vec_nodes;
+  for (const pair<long int, GraphNode>& id_and_node : nodes_) {
     vec_nodes.push_back(id_and_node);
   }
   return vec_nodes;
 }
 
-vector<int> Graph::getJunctions() const {
-  vector<int> junctions;
-  int max_degree = edge_container_.getMaxDegree();
-  for (int degree = 3; degree <= max_degree; ++degree) {
+vector<long int> Graph::getJunctions() const {
+  vector<long int> junctions;
+  long int max_degree = edge_container_.getMaxDegree();
+  for (long int degree = 3; degree <= max_degree; ++degree) {
     vector<long int> vertices = edge_container_.getVerticesDegree(degree);
     junctions.insert(junctions.end(), vertices.begin(), vertices.end());
   }
@@ -117,7 +117,7 @@ void Graph::clearNodes() { nodes_.clear(); }
 
 void Graph::copyNodes(Graph& graph) {
   assert(nodes_.size() == 0);
-  for (const pair<int, GraphNode>& id_and_node : graph.nodes_) {
+  for (const pair<long int, GraphNode>& id_and_node : graph.nodes_) {
     this->nodes_[id_and_node.first] = id_and_node.second;
   }
 }
@@ -126,14 +126,14 @@ void Graph::calcId_() {
   auto nodes = getNodes();
   sort(nodes.begin(), nodes.end(), cmpVertNodePair);
   string struct_Id_temp = "";
-  for (const pair<int, GraphNode>& id_and_node : nodes) {
+  for (const pair<long int, GraphNode>& id_and_node : nodes) {
     struct_Id_temp.append(id_and_node.second.getStringId());
   }
   id_ = struct_Id_temp;
   return;
 }
 
-int Graph::getDegree(int vertex) const {
+long int Graph::getDegree(long int vertex) const {
   if (edge_container_.vertexExist(vertex)) {
     return edge_container_.getDegree(vertex);
   }
@@ -145,7 +145,7 @@ int Graph::getDegree(int vertex) const {
       "not defined.");
 }
 
-bool Graph::vertexExist(int vertex) const {
+bool Graph::vertexExist(long int vertex) const {
   if (edge_container_.vertexExist(vertex)) {
     return true;
   }
@@ -163,15 +163,15 @@ vector<long int> Graph::getVertices() const { return edge_container_.getVertices
 
 ostream& operator<<(ostream& os, const Graph graph) {
   os << "Graph" << endl;
-  for (const pair<int, GraphNode>& id_and_node : graph.nodes_) {
+  for (const pair<long int, GraphNode>& id_and_node : graph.nodes_) {
     os << "Node " << id_and_node.first << endl;
     os << id_and_node.second << endl;
   }
   return os;
 }
 
-bool cmpVertNodePair(const pair<int, GraphNode>& id_and_node1,
-                     const pair<int, GraphNode>& id_and_node2) {
+bool cmpVertNodePair(const pair<long int, GraphNode>& id_and_node1,
+                     const pair<long int, GraphNode>& id_and_node2) {
   string str1_Id = id_and_node1.second.getStringId();
   return str1_Id.compare(id_and_node2.second.getStringId()) < 0;
 }
