+++ src/libtools/graphvisitor.cc
@@ -33,13 +33,13 @@ class GraphNode;
 
 bool GraphVisitor::queEmpty() const { return true; }
 
-void GraphVisitor::exploreNode(pair<int, GraphNode>& vertex_and_node,
+void GraphVisitor::exploreNode(pair<long int, GraphNode>& vertex_and_node,
                                Graph& graph, Edge edge) {
   explored_.insert(vertex_and_node.first);
 }
 
-vector<int> GraphVisitor::getUnexploredVertex(const Edge edge) const {
-  vector<int> unexp_vert;
+vector<long int> GraphVisitor::getUnexploredVertex(const Edge edge) const {
+  vector<long int> unexp_vert;
   if (explored_.count(edge.getEndPoint1()) == 0) {
     unexp_vert.push_back(edge.getEndPoint1());
   }
@@ -49,20 +49,20 @@ vector<int> GraphVisitor::getUnexploredVertex(const Edge edge) const {
   return unexp_vert;
 }
 
-bool GraphVisitor::vertexExplored(const int vertex) const {
+bool GraphVisitor::vertexExplored(const long int vertex) const {
   return explored_.count(vertex) == 1;
 }
 
 void GraphVisitor::initialize(Graph& graph) {
   vector<Edge> neigh_eds = graph.getNeighEdges(startingVertex_);
   GraphNode graph_node = graph.getNode(startingVertex_);
-  pair<int, GraphNode> vertex_and_graph_node(startingVertex_, graph_node);
+  pair<long int, GraphNode> vertex_and_graph_node(startingVertex_, graph_node);
   exploreNode(vertex_and_graph_node, graph);
   addEdges_(graph, startingVertex_);
 }
 
 void GraphVisitor::exec(Graph& graph, Edge edge) {
-  vector<int> unexp_vert = getUnexploredVertex(edge);
+  vector<long int> unexp_vert = getUnexploredVertex(edge);
   // If no vertices are return than just ignore it means the same
   // vertex was explored from a different direction
   if (!unexp_vert.size()) {
@@ -75,7 +75,7 @@ void GraphVisitor::exec(Graph& graph, Edge edge) {
         " did you set the starting node");
   }
 
-  pair<int, GraphNode> vertex_and_node(unexp_vert.at(0),
+  pair<long int, GraphNode> vertex_and_node(unexp_vert.at(0),
                                        graph.getNode(unexp_vert.at(0)));
 
   exploreNode(vertex_and_node, graph, edge);
@@ -86,7 +86,7 @@ Edge GraphVisitor::nextEdge(Graph graph) {
   // Get the edge and at the same time remove it from whatever queue it is in
 
   Edge edge = getEdge_(graph);
-  vector<int> unexplored_vertices = getUnexploredVertex(edge);
+  vector<long int> unexplored_vertices = getUnexploredVertex(edge);
   // Do not add neighboring edges if they belong to a vertex that has already
   // been explored because they will have already been added
   if (unexplored_vertices.size()) {
@@ -95,7 +95,7 @@ Edge GraphVisitor::nextEdge(Graph graph) {
   return edge;
 }
 
-set<int> GraphVisitor::getExploredVertices() const { return explored_; }
+set<long int> GraphVisitor::getExploredVertices() const { return explored_; }
 
 }  // namespace tools
 }  // namespace votca
